Refactored respondables and response actions
Respondables now have a class heirarchy, and response actions live as
private classes within the respondables where they are used. This should
make it easier to see which respondables are available to base your
views upon, and make it more obvious that response actions are not to be
used externally.
