--- lib/aldous/headable.rb
@@ -1,9 +1,0 @@
-module Aldous
-  module Headable
-    include Respondable
-
-    def action(controller)
-      ::Aldous::ResponseAction::Head.new(controller)
-    end
-  end
-end
