--- lib/aldous/respondable.rb
@@ -1,14 +1,0 @@
-module Aldous
-  module Respondable
-    attr_reader :result, :view_context
-
-    def initialize(result, view_context)
-      @result = result
-      @view_context = view_context
-    end
-
-    def build_result(data_hash)
-      result.class.new(data_hash)
-    end
-  end
-end
