+++ lib/aldous/respondable/request_http_basic_authentication.rb
@@ -1,0 +1,23 @@
+module Aldous
+  module Respondable
+    class RequestHttpBasicAuthentication < Base
+      def action(controller)
+        RequestHttpBasicAuthenticationAction.new(controller)
+      end
+
+      private
+
+      class RequestHttpBasicAuthenticationAction
+        attr_reader :controller
+
+        def initialize(controller)
+          @controller = controller
+        end
+
+        def execute(response_status = nil)
+          controller.request_http_basic_authentication
+        end
+      end
+    end
+  end
+end
