+++ lib/aldous/respondable/send_data.rb
@@ -1,0 +1,36 @@
+require 'aldous/respondable/base'
+
+module Aldous
+  module Respondable
+    class SendData < Base
+      def action(controller)
+        SendDataAction.new(data, options, controller, result)
+      end
+
+      def data
+        raise RuntimeError.new("SendData objects must define a 'data' method")
+      end
+
+      def options
+        raise RuntimeError.new("SendData objects must define an 'options' method")
+      end
+
+      private
+
+      class SendDataAction
+        attr_reader :controller, :result, :data, :options
+
+        def initialize(data, options, controller, result)
+          @controller = controller
+          @result = result
+          @data = data
+          @options = options
+        end
+
+        def execute(response_status = nil)
+          controller.send_data data, options
+        end
+      end
+    end
+  end
+end
