--- lib/aldous/response_action/request_http_basic_authentication.rb
@@ -1,16 +1,0 @@
-module Aldous
-  module ResponseAction
-    class RequestHttpBasicAuthentication
-      attr_reader :controller
-
-      def initialize(controller)
-        @controller = controller
-      end
-
-      def execute(response_status = nil)
-        controller.request_http_basic_authentication
-      end
-    end
-  end
-end
-
