+++ spec/aldous/respondable/base_spec.rb
@@ -1,14 +1,10 @@
-RSpec.describe Aldous::Respondable do
-  class Dummy
-    include Aldous::Respondable
-  end
-
+RSpec.describe Aldous::Respondable::Base do
   let(:result) {double("result")}
   let(:view_context) {double("view context")}
 
   describe "::new" do
     it "object can be instantiated with the right params" do
-      expect{Dummy.new(result, view_context)}.to_not raise_error
+      expect{described_class.new(result, view_context)}.to_not raise_error
     end
   end
 end
